+++ tests/cases/Help.test.js
@@ -6,31 +6,31 @@ describe('Case: Help', () => {
   function setupBattlecry() {
     battlecry.load(`${__dirname}/../fixtures/battlecry`);
     battlecry.singleHelp = jest.fn();
-    battlecry.generators.unit.help = jest.fn();
+    battlecry.strategies.unit.help = jest.fn();
   }
 
-  describe('given a generator', () => {
+  describe('given a strategy', () => {
     beforeEach(() => {
       battlecry = new Battlecry(['node', 'cry', '--help', 'unit']);
       setupBattlecry();
     });
 
-    it('shows help for the requested generator', () => {
+    it('shows help for the requested strategy', () => {
       battlecry.play();
-      expect(battlecry.generators.unit.help).not.toHaveBeenCalled();
+      expect(battlecry.strategies.unit.help).not.toHaveBeenCalled();
       expect(battlecry.singleHelp).toHaveBeenCalled();
     });
   });
 
-  describe('without a specific generator', () => {
+  describe('without a specific strategy', () => {
     beforeEach(() => {
       battlecry = new Battlecry(['node', 'cry', '--help']);
       setupBattlecry();
     });
 
-    it('shows help for all generators', () => {
+    it('shows help for all strategies', () => {
       battlecry.play();
-      expect(battlecry.generators.unit.help).toHaveBeenCalled();
+      expect(battlecry.strategies.unit.help).toHaveBeenCalled();
       expect(battlecry.singleHelp).not.toHaveBeenCalled();
     });
   });