+++ tests/decorators/option.test.js
@@ -2,15 +2,15 @@ import { command } from 'battlecry';
 
 describe('@option', () => {
   describe('without details', () => {
-    class TestGenerator {
+    class TestStrategy {
       @command
       @command.option('path')
       test() {}
     }
 
     it('registers the option', () => {
-      const generator = new TestGenerator();
-      expect(generator.config).toEqual({
+      const strategy = new TestStrategy();
+      expect(strategy.config).toEqual({
         test: {
           options: {
             path: {}
@@ -22,15 +22,15 @@ describe('@option', () => {
 });
 
 describe('with details', () => {
-  class TestGenerator {
+  class TestStrategy {
     @command
     @command.option('path', { alias: 'j' })
     test() {}
   }
 
   it('registers the option with details', () => {
-    const generator = new TestGenerator();
-    expect(generator.config).toEqual({
+    const strategy = new TestStrategy();
+    expect(strategy.config).toEqual({
       test: {
         options: {
           path: {